requirements.txt
django==2.2
psycopg2==2.8.3
djangorestframework==3.10.3 
